Change the type of prime

+++ src/types.h
@@ -18,6 +18,6 @@ using ExpansionBlockSet = ExpansionBlock[EXPANSION_BLOCK_COUNT_IN_SET];
 
 // Compression phase types
 using HashBlock = std::bitset<HASH_OUTPUT_SIZE>;
-using Prime = unsigned int;
+using Prime = unsigned long int;
 
 #endif //BITCOIN_ENCRYPTION_TYPES_H
